requirements.txt
# If not already create a python enviroment
    # python3 -m venv <env name>
    # source <env name>/bin/activate
# Before this Run: 
    # pip3 install --upgrade pip setuptools wheel
# Install with: 
    # pip3 install -r requirements.txt

black
flake8
requests
pytz
pydantic
pymongo
playwright   # see notes below
python-dotenv
Flask
gunicorn


# For playwright on server:
# sudo apt update
# sudo apt install -y \
#     libnss3 \
#     libatk-bridge2.0-0t64 \
#     libx11-xcb1 \
#     libxcomposite1 \
#     libxdamage1 \
#     libxrandr2 \
#     libasound2t64 \
#     libgbm1 \
#     libgtk-3-0t64 \
#     libgtk-4-1 \
#     libpangocairo-1.0-0 \
#     libcups2t64 \
#     libdrm2 \
#     libxss1 \
#     libwoff1 \
#     libharfbuzz0b \
#     libharfbuzz-icu0 \
#     libpango-1.0-0 \
#     libatspi2.0-0t64 \
#     libxcb1 \
#     libxext6 \
#     libgraphene-1.0-0 \
#     libvpx9 \
#     libevent-2.1-7t64 \
#     libopus0 \
#     gstreamer1.0-plugins-base \
#     gstreamer1.0-plugins-good \
#     gstreamer1.0-plugins-bad \
#     gstreamer1.0-plugins-ugly \
#     gstreamer1.0-libav \
#     gstreamer1.0-tools \
#     libflite1 \
#     libwebpdemux2 \
#     libwebpmux3 \
#     libavif16 \
#     libenchant-2-2 \
#     libsecret-1-0 \
#     libhyphen0 \
#     libmanette-0.2-0 \
#     libx264-dev

# Then run:
# python -m playwright install


# For playwright on mac:
# python -m playwright install
